refactor: move subroutines to pkg to allow for reuse
