+++ .testcoverage.yml
@@ -1,6 +1,6 @@
 exclude:
   paths:
-    - ^internal/subroutines/mocks # exclude generated mock files
+    - ^pkg/subroutines/mocks # exclude generated mock files
     - api/v1alpha1 # skipping generated files and crd type definitions
     - main\.go$ # skip covering main.go
